reorg and hoist view model
