+++ app/src/main/java/com/shinjiindustrial/portmapper/MainActivity.kt
@@ -165,8 +165,10 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import org.fourthline.cling.model.meta.RemoteDevice
+import java.com.shinjiindustrial.portmapper.PortMapperNavGraph
 import java.com.shinjiindustrial.portmapper.PortViewModel
 import java.com.shinjiindustrial.portmapper.SettingsViewModel
+import java.com.shinjiindustrial.portmapper.ThemeUiState
 import java.net.InetAddress
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
@@ -297,503 +299,24 @@ class MainActivity : ComponentActivity() {
         portViewModel.start()
 
         setContent {
-            MainContent()
+            PortMapperMainContent(portViewModel)
         }
     }
 
     @Composable
-    fun MainContent() {
+    fun PortMapperMainContent(portViewModel : PortViewModel) {
         val themeState by settingsViewModel.uiState.collectAsStateWithLifecycle()
         MyApplicationTheme(themeState) {
             Surface(
                 modifier = Modifier.fillMaxSize(),
                 color = MaterialTheme.colorScheme.background
             ) {
-                AppNavigator()
+                PortMapperNavGraph(portViewModel, themeState)
             }
         }
     }
 
 
-    @OptIn(ExperimentalAnimationApi::class)
-    @Composable
-    fun AppNavigator() {
-        val navController = rememberNavController()
-
-        NavHost(navController = navController, startDestination = "main_screen") {
-            composable(
-                "main_screen",
-                exitTransition = {
-                    fadeOut(animationSpec = tween(300))
-                },
-                popEnterTransition = {
-                    //this is whats used when going back from "create rule"
-                    slideInVertically(
-                        initialOffsetY = { it / 2 }, animationSpec = tween(
-                            durationMillis = 200, easing = FastOutSlowInEasing
-                        )
-                    ) + fadeIn(animationSpec = tween(400))
-                },
-                popExitTransition = {
-                    fadeOut(animationSpec = tween(300))
-                },
-            )
-            {
-                MainScreen(navController = navController)
-            }
-            composable(
-                "full_screen_dialog?description={description}&internalIp={internalIp}&internalRange={internalRange}&externalIp={externalIp}&externalRange={externalRange}&protocol={protocol}&leaseDuration={leaseDuration}&enabled={enabled}&autorenew={autorenew}",
-                arguments = listOf(
-                    navArgument("description") {
-                        nullable = true; type = NavType.StringType
-                    }, // only if nullable (or default) are they optional
-                    navArgument("internalIp") { nullable = true; type = NavType.StringType },
-                    navArgument("internalRange") { nullable = true; type = NavType.StringType },
-                    navArgument("externalIp") { nullable = true; type = NavType.StringType },
-                    navArgument("externalRange") { nullable = true; type = NavType.StringType },
-                    navArgument("protocol") { nullable = true; type = NavType.StringType },
-                    navArgument("leaseDuration") { nullable = true; type = NavType.StringType },
-                    navArgument("enabled") { type = NavType.BoolType; defaultValue = false },
-                    navArgument("autorenew") { type = NavType.BoolType; defaultValue = false },
-                ),
-                popExitTransition = {
-
-                    slideOutVertically(
-                        targetOffsetY = { it / 2 }, animationSpec = tween(
-                            durationMillis = 200, easing = FastOutSlowInEasing
-                        )
-                    ) + fadeOut(animationSpec = tween(100))
-                },
-                exitTransition = {
-                    slideOutVertically(
-                        targetOffsetY = { it }, animationSpec = tween(
-                            durationMillis = 5000, easing = FastOutSlowInEasing
-                        )
-                    )
-                },
-                enterTransition = {
-                    slideInVertically(
-                        initialOffsetY = { it }, animationSpec = tween(
-                            durationMillis = 200, easing = FastOutSlowInEasing
-                        )
-                    ) + fadeIn(animationSpec = tween(400))
-                },
-                popEnterTransition = {
-                    fadeIn(animationSpec = tween(150))
-                }
-            )
-            { backStackEntry ->
-                val arguments = backStackEntry.arguments
-                val desc = arguments?.getString("description")
-                val internalIp = arguments?.getString("internalIp")
-                val internalRange = arguments?.getString("internalRange")
-                val externalIp = arguments?.getString("externalIp")
-                val externalRange = arguments?.getString("externalRange")
-                val protocol = arguments?.getString("protocol")
-                val leaseDuration = arguments?.getString("leaseDuration")
-                val enabled = arguments?.getBoolean("enabled")
-                val autorenew = arguments?.getBoolean("autorenew")
-
-                var portMappingUserInputToEdit: PortMappingUserInput? = null
-                if (desc != null) {
-                    portMappingUserInputToEdit = PortMappingUserInput(
-                        desc,
-                        internalIp!!,
-                        internalRange!!,
-                        externalIp!!,
-                        externalRange!!,
-                        protocol!!,
-                        leaseDuration!!,
-                        enabled!!,
-                        autorenew!!
-                    )
-                }
-
-                RuleCreationDialog(
-                    navController = navController,
-                    portViewModel,
-                    ruleToEdit = portMappingUserInputToEdit
-                )
-            }
-        }
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Composable
-    fun MainScreen(navController: NavHostController) {
-        // A surface container using the 'background' color from the theme
-        val searchStartedRecentlyAndNothingFoundYet by portViewModel.searchStartedRecentlyAndNothingFoundYet.collectAsStateWithLifecycle()
-        val anyDevices by portViewModel.anyDevices.collectAsStateWithLifecycle()
-        rememberScrollState()
-        val showAboutDialogState = remember { mutableStateOf(false) }
-        val showMoreInfoDialogState = remember { mutableStateOf(false) }
-        val showAboutDialog by showAboutDialogState //mutable state binds to UI (in sense if value changes, redraw). remember says when redrawing dont discard us.
-        val inMultiSelectMode by portViewModel.inMultiSelectMode.collectAsStateWithLifecycle()
-        val selectedIds by portViewModel.selectedIds.collectAsStateWithLifecycle()
-
-        LaunchedEffect(Unit) {
-            portViewModel.events.collect { ev ->
-                if (ev is PortViewModel.UiEvent.ToastEvent) {
-                    Toast.makeText(
-                        PortForwardApplication.appContext,
-                        ev.msg,
-                        ev.duration
-                    ).show()
-                } else if (ev is PortViewModel.UiEvent.SnackBarViewLogEvent) {
-                    showSnackBarViewLog(ev.msg) // emit these
-                }
-            }
-        }
-
-        PortForwardApplication.currentSingleSelectedObject =
-            remember { mutableStateOf(null) }
-
-        if (PortForwardApplication.showContextMenu.value && PortForwardApplication.currentSingleSelectedObject.value != null) {
-            EnterContextMenu(
-                PortForwardApplication.currentSingleSelectedObject, // why is this not typed??
-                showMoreInfoDialogState,
-                navController,
-                portViewModel,
-                settingsViewModel
-            )
-        }
-
-        if (showMoreInfoDialogState.value) {
-            MoreInfoDialog(
-                portMappingWithPref = PortForwardApplication.currentSingleSelectedObject.value as PortMappingWithPref,
-                showDialog = showMoreInfoDialogState
-            )
-        }
-
-        if (showAboutDialog) {
-            AlertDialog(
-                onDismissRequest = { showAboutDialogState.value = false },
-                title = { Text("About") },
-                text = {
-
-                    Column()
-                    {
-                        var aboutString: String =
-                            PortForwardApplication.appContext.getString(R.string.about_body)
-                        val packageInfo =
-                            PortForwardApplication.appContext.packageManager.getPackageInfo(
-                                PortForwardApplication.appContext.packageName,
-                                0
-                            )
-
-                        aboutString = String.format(aboutString, packageInfo.versionName)
-                        var spannedString: Spanned? = null
-                        if (Build.VERSION.SDK_INT >= 24) {
-                            spannedString =
-                                Html.fromHtml(aboutString, Html.FROM_HTML_MODE_LEGACY)
-                        } else {
-                            spannedString = Html.fromHtml(aboutString)
-                        }
-
-                        // compose doesn't allow spannable nor link movement method
-                        AndroidView(factory = { context ->
-                            TextView(context).apply {
-                                setTextColor(AdditionalColors.TextColor.toArgb())
-                                textSize = 16f
-                                text = spannedString
-                                movementMethod = LinkMovementMethod.getInstance()
-                                // you can apply other TextView properties here
-                            }
-                        })
-
-                        //Text(aboutString)
-                    }
-
-
-                },
-                confirmButton = {
-                    Button(onClick = { showAboutDialogState.value = false }) {
-                        Text("OK")
-                    }
-                })
-        }
-
-
-        OurSnackbarHostState = remember { SnackbarHostState() }
-        rememberCoroutineScope()
-        val coroutineScope: CoroutineScope = rememberCoroutineScope()
-
-        var openBottomSheet by remember { mutableStateOf(false) }
-        val bottomSheetState =
-            rememberModalBottomSheetState(true)
-
-        if (openBottomSheet) {
-            ModalBottomSheet(
-                onDismissRequest = { openBottomSheet = false },
-                sheetState = bottomSheetState,
-                shape = RoundedCornerShape(topStart = 24.dp, topEnd = 24.dp),
-                tonalElevation = 24.dp,
-                containerColor = MaterialTheme.colorScheme.surfaceColorAtElevation(2.dp),
-                content = {
-                    // This is what will be shown in the bottom sheet
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp)
-                    ) {
-                        BottomSheetSortBy(portViewModel)
-                    }
-                }
-            )
-//
-//            LaunchedEffect(Unit) {
-//                bottomSheetState.expand()
-//            }
-        }
-
-        Scaffold(
-
-            snackbarHost = {
-                SnackbarHost(OurSnackbarHostState!!) { data ->
-                    // custom snackbar with the custom colors
-                    Snackbar(
-                        data,
-                        actionColor = AdditionalColors.PrimaryDarkerBlue
-                        // according to https://m2.material.io/design/color/dark-theme.html
-                        // light snackbar in darkmode is good.
-//                                    containerColor = AdditionalColors.CardContainerColor,
-//                                    contentColor = AdditionalColors.TextColor
-//                                    //contentColor = ...,
-                    )
-                }
-            },
-            floatingActionButton = {
-
-                if (anyDevices) {
-                    FloatingActionButton(
-                        // uses MaterialTheme.colorScheme.secondaryContainer
-                        containerColor = MaterialTheme.colorScheme.secondaryContainer, //todo revert to secondar
-                        onClick = {
-                            //showAddRuleDialogState.value = true
-                            navController.navigate("full_screen_dialog")
-
-                            //this works
-
-                        }) {
-                        Icon(
-                            Icons.Default.Add,
-                            contentDescription = "Localized description",
-                            tint = AdditionalColors.TextColor,
-                        )
-                    }
-                }
-            },
-            topBar = {
-                TopAppBar(
-                    navigationIcon = {
-                        if (inMultiSelectMode) {
-                            IconButton(onClick = {
-                                portViewModel.clearSelection()
-                            })
-                            {
-                                Icon(Icons.Filled.ArrowBack, contentDescription = "Back")
-                            }
-                        }
-                    },
-//                                modifier = Modifier.height(40.dp),
-                    colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
-                        containerColor = AdditionalColors.TopAppBarColor
-                    ),
-                    //colors = TopAppBarDefaults.smallTopAppBarColors(containerColor = MaterialTheme.colorScheme.secondary),// change the height here
-                    title = {
-                        val title =
-                            if (inMultiSelectMode) "${selectedIds.size} Selected" else "PortMapper"
-                        Text(
-                            text = title,
-                            color = AdditionalColors.TextColorStrong,
-                            fontWeight = FontWeight.Normal
-                        )
-                    },
-                    actions = {
-
-                        if (inMultiSelectMode) {
-                            IconButton(onClick = {
-                                if (inMultiSelectMode) {
-                                    portViewModel.deleteAll(selectedIds)
-                                } else {
-                                    portViewModel.deleteAll()
-                                }
-                            })
-                            {
-                                Icon(Icons.Default.Delete, contentDescription = "Delete")
-                            }
-                        } else {
-                            IconButton(onClick = {
-                                coroutineScope.launch {
-                                    openBottomSheet = true
-                                    bottomSheetState.show()
-                                }
-                            })
-                            {
-                                Icon(Icons.Default.Sort, contentDescription = "Sort")
-                            }
-                        }
-
-                        OverflowMenu(
-                            showAboutDialogState, portViewModel
-                        )
-                    }
-                )
-            },
-            content = { it ->
-
-                // TODO can revisit this - if we already know we have some kind of device
-                //   then we can forget the main loading circle and show this until we
-                //   finish enumerating our first device (or have a flow which is like
-                //   combine (deviceNotFound, anyDeviceStillBeingSearched))
-                // still the issue of if search finishes really fast then it can hang if
-                //   you set isRefreshing to false too fast (can fix by having an arbitrary delay)
-
-                val state = rememberPullToRefreshState()
-                val scope = rememberCoroutineScope()
-                val isRefreshing = remember { mutableStateOf(false) }
-                fun refresh() = scope.launch {
-
-                    state.animateToHidden()
-                    portViewModel.fullRefresh()
-                }
-
-                PullToRefreshBox(
-                    isRefreshing = isRefreshing.value,
-                    onRefresh = ::refresh,
-                    modifier = Modifier.padding(it),
-                    state = state,
-                    indicator = {
-                        // dont show after initially pulling down
-                        if (state.distanceFraction > 0f) {
-                            Indicator(
-                                state = state,
-                                isRefreshing = isRefreshing.value,
-                                modifier = Modifier.align(Alignment.TopCenter)
-                            )
-
-                        }
-                    }
-                ) {
-                    var heightPx by remember { mutableIntStateOf(0) }
-                    val boxHeight = with(LocalDensity.current) { heightPx.toDp() }
-                    Column(
-                        Modifier
-                            .onSizeChanged { heightPx = it.height }
-                            .fillMaxHeight()
-                            .fillMaxWidth()
-                    ) {
-                        if (searchStartedRecentlyAndNothingFoundYet) {
-                            val offset = boxHeight * 0.28f
-                            LoadingIcon(
-                                "Searching for devices",
-                                Modifier.offset(y = offset)
-                            )
-                        } else if (!anyDevices) {
-                            val offset = boxHeight * 0.28f
-                            Column(modifier = Modifier.offset(y = offset))
-                            {
-                                Text(
-                                    "No UPnP enabled internet gateway devices found",
-                                    modifier = Modifier
-                                        .padding(0.dp, 10.dp)
-                                        .align(Alignment.CenterHorizontally),
-                                    textAlign = TextAlign.Center
-                                )
-
-                                // these can go through vm -> repo -> client
-                                val interfacesUsedInSearch =
-                                    portViewModel.getInterfacesUsedInSearch()
-                                val anyInterfaces =
-                                    portViewModel.isInitialized()
-                                if (!anyInterfaces) {
-                                    Text(
-                                        "No valid interfaces",
-                                        modifier = Modifier
-                                            .padding(0.dp, 10.dp)
-                                            .align(Alignment.CenterHorizontally),
-                                        textAlign = TextAlign.Center
-                                    )
-                                } else {
-
-                                    Text(
-                                        "Interfaces Searched:",
-                                        modifier = Modifier
-                                            .padding(0.dp, 10.dp, 0.dp, 0.dp)
-                                            .align(Alignment.CenterHorizontally),
-                                        textAlign = TextAlign.Center
-                                    )
-
-
-                                    for (interfaceUsed in interfacesUsedInSearch!!) {
-                                        Text(
-                                            "${interfaceUsed.networkInterface.displayName} (${interfaceUsed.networkType.networkTypeString.lowercase()})",
-                                            modifier = Modifier
-                                                .padding(0.dp, 0.dp)
-                                                .align(Alignment.CenterHorizontally),
-                                            textAlign = TextAlign.Center
-                                        )
-                                    }
-                                }
-
-                                // if no wifi
-                                if (!interfacesUsedInSearch.any { it.networkType == NetworkType.WIFI }
-                                ) {
-                                    Text(
-                                        "Enable WiFi and retry",
-                                        modifier = Modifier
-                                            .padding(0.dp, 20.dp, 0.dp, 10.dp)
-                                            .align(Alignment.CenterHorizontally),
-                                        textAlign = TextAlign.Center
-                                    )
-                                } else {
-                                    Text(
-                                        "Ensure a valid UPnP enabled internet gateway device is on the network and retry",
-                                        modifier = Modifier
-                                            .padding(0.dp, 20.dp, 0.dp, 10.dp)
-                                            .align(Alignment.CenterHorizontally),
-                                        textAlign = TextAlign.Center
-                                    )
-                                }
-                                Button(
-                                    onClick = {
-                                        portViewModel.initialize(
-                                            PortForwardApplication.appContext,
-                                            true
-                                        )
-                                        portViewModel.fullRefresh()
-                                    },
-                                    modifier = Modifier
-                                        .padding(0.dp, 10.dp)
-                                        .align(Alignment.CenterHorizontally),
-                                    shape = RoundedCornerShape(4),
-                                ) {
-                                    Text(
-                                        "Retry",
-                                        color = AdditionalColors.TextColorStrong
-                                    )
-                                }
-                            }
-                        } else {
-                            val uiState by portViewModel.uiState.collectAsStateWithLifecycle()
-                            val selectedIds by portViewModel.selectedIds.collectAsStateWithLifecycle()
-                            val isInMultiSelectMode by portViewModel.inMultiSelectMode.collectAsStateWithLifecycle()
-                            PortMappingContent(
-                                uiState,
-                                isInMultiSelectMode,
-                                portViewModel::toggle,
-                                selectedIds
-                            )
-                        }
-                    }
-
-                }
-
-            }
-        )
-
-
-    }
 
 }
 
@@ -821,7 +344,7 @@ fun EnterContextMenu(
     showMoreInfoDialog: MutableState<Boolean>,
     navController: NavHostController,
     portViewModel: PortViewModel,
-    settingsViewModel: SettingsViewModel
+    themeState: ThemeUiState
 ) {
     if (singleSelectedItem.value == null) {
         return
@@ -834,7 +357,6 @@ fun EnterContextMenu(
         usePlatformDefaultWidth = false,
         decorFitsSystemWindows = true,
     )
-    val themeState by settingsViewModel.uiState.collectAsStateWithLifecycle()
     MyApplicationTheme(themeState) {
         Dialog(
             onDismissRequest = { PortForwardApplication.showContextMenu.value = false },
